+++ R/drop-rows.R
@@ -10,10 +10,12 @@
 #'}
 #' @export
 dropRows <- function (dataset, expr) {
-    ## Construct a ZCL delete payload, then POST it
-    payload <- list(command="delete",
-        filter=zcl(expr))
+    if (!length(expr)) {
+        ## Could be NULL or like NULL == something, which is logical(0)
+        halt("Invalid expression: ", deparse(match.call()$expr)[1])
+    }
+    payload <- list(command="delete", filter=zcl(expr))
     out <- crPOST(shojiURL(dataset, "fragments", "table"), body=toJSON(payload))
-    dropCache(self(dataset))
+    dropCache(self(dataset)) ## Could do a more surgical cache drop, but this is safe
     return(dataset)
 }