+++ tests/testthat/test-crunchbox.R
@@ -7,13 +7,59 @@ test_that("Box size limit check", {
     expect_true(boxTooBig(100, 6))
 })
 
+test_that("Embed URL", {
+    expect_identical(boxdataToWidgetURL("http://cf.example/d/stuff/1a1577c91fbb2c1cbd3800e181188508/dataset.json"),
+        "//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/")
+    expect_identical(boxdataToWidgetURL("//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/"),
+        "//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/")
+})
+
+test_that("Iframe code (prints and returns invisibly)", {
+    ## Use the testthat version of expect_output because the crunch version
+    ## calls print explicitly
+    testthat::expect_output(
+        expect_identical(embedCrunchBox("http://cf.example/d/stuff/1a1577c91fbb2c1cbd3800e181188508/dataset.json"),
+            '<iframe src="//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/" width="600" height="480" style="border: 1px solid #d3d3d3;"></iframe>'),
+        '<iframe src="//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/" width="600" height="480" style="border: 1px solid #d3d3d3;"></iframe>',
+        fixed=TRUE)
+})
+
+iframe_with_logo <- '<figure style="text-align: left;" class="content-list-component image">
+    <img src="//s.crunch.io/public/branding/example.gif" style="height:auto; width:200px; margin-left:-4px"></img>
+    <iframe src="//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/" width="600" height="480" style="border: 1px solid #d3d3d3;"></iframe>
+</figure>'
+
+iframe_with_title <- '<figure style="text-align: left;" class="content-list-component image">
+    <div style="padding-bottom: 12px">
+        <span style="font-size: 18px; color: #444444; line-height: 1;">Example title here</span>
+    </div>
+    <iframe src="//s.crunch.io/widget/index.html#/ds/1a1577c91fbb2c1cbd3800e181188508/" width="600" height="480" style="border: 1px solid #d3d3d3;"></iframe>
+</figure>'
+
+
+test_that("Iframe code with logo", {
+    testthat::expect_output(
+        expect_identical(embedCrunchBox("http://cf.example/d/stuff/1a1577c91fbb2c1cbd3800e181188508/dataset.json", logo="//s.crunch.io/public/branding/example.gif"),
+            iframe_with_logo),
+        iframe_with_logo,
+        fixed=TRUE)
+})
+
+test_that("Iframe code with title", {
+    testthat::expect_output(
+        expect_identical(embedCrunchBox("http://cf.example/d/stuff/1a1577c91fbb2c1cbd3800e181188508/dataset.json", title="Example title here"),
+            iframe_with_title),
+        iframe_with_title,
+        fixed=TRUE)
+})
+
 with_mock_HTTP({
     ds <- loadDataset("test ds")
     ds3 <- loadDataset("ECON.sav")
 
     test_that("preCrunchBoxCheck does not error", {
         expect_output(preCrunchBoxCheck(ds),
-            "We recommend using only categorical and multiple_response variables. These 4 variables are not")
+            "We recommend using only categorical and multiple_response variables. These 4 variables have an unsupported type")
     })
 
     test_that("Basic box", {
@@ -77,5 +123,29 @@ with_mock_HTTP({
             expect_error(crunchBox(ds[2]),
                 "1 variable and 2 filters results in too many cubes")
         })
+    })
+})
+
+with_test_authentication({
+    testdf <- as.data.frame(sapply(letters[1:8], function (x) df$v4, simplify=FALSE))
+    testdf$cat <- as.factor(letters[1:10])
+    testdf$num <- 1
+    ds <- newDataset(testdf)
+    ds$mr <- makeMR(ds[letters[1:8]],
+        name="Excessively long variable name to trigger the check for length",
+        selections="B")
+    names(subvariables(ds$mr))[1] <- "Another really really long name to check for the same for subvariables"
+    names(categories(ds$cat))[2] <- "Extra long category name because we check those too"
+
+    test_that("check box catches the various cases", {
+        expect_output(preCrunchBoxCheck(ds),
+            "Shorter variable names will display in the menus better. This variable has a name longer than 40 characters")
+    })
+
+    weight(ds) <- ds$num
+    hiddenVariables(ds) <- "num"
+    filters(ds)[["A filter"]] <- ds$cat == "d"
+    test_that("We can make a box", {
+        expect_true(is.character(crunchBox(ds)))
     })
 })